upstream-sync: bootstrap — session-start + OpenCode consolidation (from v6.0.3)
session-start:
- adopt: drop the legacy ~/.config/superflowers/skills warning (that path never
  existed for this fork) and add `| cat` to the 3 printf outputs (SIGPIPE/buffer fix)
- kept: our Copilot CLI platform detection + superflowers rebrand

OpenCode plugin consolidation (fork had two divergent plugin files):
- .opencode/plugins/superflowers.js is now the single, up-to-date plugin =
  upstream v6.0.3 (messages.transform, fixes token bloat #750 + multi-system-msg
  #894; vendor-neutral tool mapping), rebranded. package.json main already points here.
- removed the stale duplicate .opencode/plugins/superpowers.js
- tests/opencode/{setup.sh,test-plugin-loading.sh}: adopt v6.0.3 + rebrand to
  reference superflowers.js
- .opencode/INSTALL.md + docs/README.opencode.md: adopt v6.0.3 + rebrand
  (install URL now florian23/superflowers)

Static checks: node --check (superflowers.js) + bash -n (session-start, tests) pass.
