+++ tests/opencode/setup.sh
@@ -14,26 +14,26 @@ export XDG_CONFIG_HOME="$TEST_HOME/.config"
 export OPENCODE_CONFIG_DIR="$TEST_HOME/.config/opencode"
 
 # Standard install layout:
-#   $OPENCODE_CONFIG_DIR/superpowers/             ← package root
-#   $OPENCODE_CONFIG_DIR/superpowers/skills/      ← skills dir (../../skills from plugin)
-#   $OPENCODE_CONFIG_DIR/superpowers/.opencode/plugins/superpowers.js ← plugin file
-#   $OPENCODE_CONFIG_DIR/plugins/superpowers.js   ← symlink OpenCode reads
+#   $OPENCODE_CONFIG_DIR/superflowers/             ← package root
+#   $OPENCODE_CONFIG_DIR/superflowers/skills/      ← skills dir (../../skills from plugin)
+#   $OPENCODE_CONFIG_DIR/superflowers/.opencode/plugins/superflowers.js ← plugin file
+#   $OPENCODE_CONFIG_DIR/plugins/superflowers.js   ← symlink OpenCode reads
 
-SUPERPOWERS_DIR="$OPENCODE_CONFIG_DIR/superpowers"
-SUPERPOWERS_SKILLS_DIR="$SUPERPOWERS_DIR/skills"
-SUPERPOWERS_PLUGIN_FILE="$SUPERPOWERS_DIR/.opencode/plugins/superpowers.js"
+SUPERFLOWERS_DIR="$OPENCODE_CONFIG_DIR/superflowers"
+SUPERFLOWERS_SKILLS_DIR="$SUPERFLOWERS_DIR/skills"
+SUPERFLOWERS_PLUGIN_FILE="$SUPERFLOWERS_DIR/.opencode/plugins/superflowers.js"
 
 # Install skills
-mkdir -p "$SUPERPOWERS_DIR"
-cp -r "$REPO_ROOT/skills" "$SUPERPOWERS_DIR/"
+mkdir -p "$SUPERFLOWERS_DIR"
+cp -r "$REPO_ROOT/skills" "$SUPERFLOWERS_DIR/"
 
 # Install plugin
-mkdir -p "$(dirname "$SUPERPOWERS_PLUGIN_FILE")"
-cp "$REPO_ROOT/.opencode/plugins/superpowers.js" "$SUPERPOWERS_PLUGIN_FILE"
+mkdir -p "$(dirname "$SUPERFLOWERS_PLUGIN_FILE")"
+cp "$REPO_ROOT/.opencode/plugins/superflowers.js" "$SUPERFLOWERS_PLUGIN_FILE"
 
 # Register plugin via symlink (what OpenCode actually reads)
 mkdir -p "$OPENCODE_CONFIG_DIR/plugins"
-ln -sf "$SUPERPOWERS_PLUGIN_FILE" "$OPENCODE_CONFIG_DIR/plugins/superpowers.js"
+ln -sf "$SUPERFLOWERS_PLUGIN_FILE" "$OPENCODE_CONFIG_DIR/plugins/superflowers.js"
 
 # Create test skills in different locations for testing
 
@@ -67,10 +67,10 @@ EOF
 
 echo "Setup complete: $TEST_HOME"
 echo "OPENCODE_CONFIG_DIR:  $OPENCODE_CONFIG_DIR"
-echo "Superpowers dir:      $SUPERPOWERS_DIR"
-echo "Skills dir:           $SUPERPOWERS_SKILLS_DIR"
-echo "Plugin file:          $SUPERPOWERS_PLUGIN_FILE"
-echo "Plugin registered at: $OPENCODE_CONFIG_DIR/plugins/superpowers.js"
+echo "Superflowers dir:      $SUPERFLOWERS_DIR"
+echo "Skills dir:           $SUPERFLOWERS_SKILLS_DIR"
+echo "Plugin file:          $SUPERFLOWERS_PLUGIN_FILE"
+echo "Plugin registered at: $OPENCODE_CONFIG_DIR/plugins/superflowers.js"
 echo "Test project at:      $TEST_HOME/test-project"
 
 # Helper function for cleanup (call from tests or trap)
@@ -83,6 +83,6 @@ cleanup_test_env() {
 # Export for use in tests
 export -f cleanup_test_env
 export REPO_ROOT
-export SUPERPOWERS_DIR
-export SUPERPOWERS_SKILLS_DIR
-export SUPERPOWERS_PLUGIN_FILE
+export SUPERFLOWERS_DIR
+export SUPERFLOWERS_SKILLS_DIR
+export SUPERFLOWERS_PLUGIN_FILE
